logic_apps/requirements.txt
# Azure SDKs
azure-identity 
azure-mgmt-resource